Add nix settings
