--- cargo/link.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-
-ln -sf ~/dotfiles/cargo/config.toml ~/.cargo/config.toml
