--- init/setup/apt.sh
@@ -1,6 +1,0 @@
-#!/bin/bash
-
-# to build ren and tree command and make vimproc
-sudo apt -y update
-sudo apt -y install g++
-sudo apt -y install zsh
