--- init/setup/brew.sh
@@ -1,49 +1,0 @@
-#!/bin/bash
-
-brew install fish
-brew install fzf
-brew install go
-brew install tmux
-brew install peco
-brew install python3
-# brew install nvim
-# install latest version
-brew install --HEAD neovim
-brew install make
-brew install jq
-brew install yq
-brew install shellcheck
-brew install tree
-brew install ren
-brew install direnv
-brew install wget
-brew install tmux-xpanes
-brew install fontforge
-brew install reattach-to-user-namespace # tmux
-brew install gnu-sed # replace \t
-brew install asciidoctor
-brew install --HEAD pyenv-virtualenv
-brew install --HEAD pyenv-virtualenvwrapper
-brew install asdf
-brew install duckdb
-brew install dotenvx/brew/dotenvx
-brew install vhs
-
-# brew install tig
-# brew install terraform
-# brew unlink terraform
-# brew install tfenv   # terraform version manager
-
-# TODO deplicate(change to cargo install. See common.sh.)
-# brew install fd
-# brew install bat
-# brew install rg
-# brew install eza
-
-# for cpp enviroment settings
-brew install llvm
-
-# aws tools
-brew tap aws/tap
-brew install aws-sam-cli
-brew install font-hack-nerd-font
