--- init/setup/frontend.sh
@@ -1,20 +1,0 @@
-#!/bin/bash
-sudo npm install -g vue-language-server
-sudo npm install -g javascript-typescript-langserver
-sudo npm install -g dockerfile-language-server-nodejs
-sudo npm install -g bash-language-server
-sudo npm install -g markdownlint
-sudo npm install -g typescript
-sudo npm install -g typescript-language-server
-sudo npm install -g @tailwindcss/language-server
-sudo npm install -g yarn
-sudo npm install -g jwt-cli
-sudo npm install -g reload
-sudo npm install -g jsondiffpatch
-sudo npm install -g uuid-cli
-sudo npm install -g lighthouse
-sudo npm install -g npm-check-updates
-sudo npm install -g pyright
-# for nvim-treesitter plugin
-sudo npm install -g tree-sitter-cli
-sudo npm install -g @biomejs/biome
