--- init/setup/yum.sh
@@ -1,5 +1,0 @@
-#!/bin/bash
-
-sudo yum update -y
-sudo yum install -y git
-sudo yum install -y gcc-c++
