+++ nix/home-manager/programs/alacritty/alacritty.toml
