+++ nix/home-manager/programs/alacritty/default.nix
@@ -1,0 +1,5 @@
+{ pkgs, lib, ... }:
+{
+  home.file.".config/alacritty/alacritty.toml".text = builtins.readFile ./alacritty.toml;
+  home.file.".config/alacritty/catppuccin-mocha.toml".text = builtins.readFile ./catppuccin-mocha.toml;
+}
