+++ nix/home-manager/programs/cargo/default.nix
@@ -1,0 +1,4 @@
+{ pkgs, lib, ... }:
+{
+  home.file.".cargo/config".text = builtins.readFile ./config.toml;
+}
