+++ nix/home-manager/programs/deno/default.nix
@@ -1,0 +1,6 @@
+{ pkgs, ... }:
+{
+  home.packages = with pkgs; [
+    deno
+  ];
+}
