+++ nix/home-manager/programs/fzf/default.nix
@@ -1,0 +1,7 @@
+{ pkgs, ... }:
+{
+  programs.fzf = {
+    enable = true;
+    tmux.enableShellIntegration = true;
+  };
+}
