+++ nix/home-manager/programs/gh/default.nix
@@ -1,0 +1,15 @@
+{ ... }:
+{
+  programs.gh = {
+    enable = true;
+    settings = {
+      git_protocol = "https";
+      aliases = {
+        co = "pr checkout";
+        pv = "pr view";
+        see = "browse";
+      };
+      editor = "nvim";
+    };
+  };
+}
