+++ nix/home-manager/programs/lazygit/default.nix
@@ -1,0 +1,6 @@
+{ pkgs, ... }:
+{
+  programs.lazygit = {
+    enable = true;
+  };
+}
