+++ nix/home-manager/programs/pet/snippet.toml
