+++ nix/home-manager/programs/tmux/default.nix
@@ -1,0 +1,8 @@
+{ pkgs, ... }:
+{
+  programs.tmux = {
+    enable = true;
+    extraConfig = builtins.readFile ./.tmux.conf;
+    shell = "${pkgs.zsh}/bin/zsh";
+  };
+}
