--- private/.gitignore
@@ -1,3 +1,0 @@
-*
-!.gitignore
-!link.sh
