--- private/link.sh
@@ -1,1 +1,0 @@
-ln -sf ~/dotfiles/private/.vimpressrc ~/.vimpressrc
