--- tmux/powerline/ssid
@@ -1,50 +1,0 @@
-#!/usr/local/bin/python3
-import re
-import subprocess
-
-
-def exec_airport():
-    res = subprocess.run(
-        [
-            "/System/Library/PrivateFrameworks/Apple80211.framework/Versions/Current/Resources/airport",
-            "-I",
-        ],
-        stdout=subprocess.PIPE,
-    )
-    return res
-
-
-class ssid:
-    def __init__(self):
-        res = exec_airport()
-        try:
-            ssid_state = res.stdout.decode().splitlines()[4]
-            match = re.search(r"state: (\w+)", ssid_state)
-            self.status = match.group(1).strip()
-        except IndexError:
-            self.status = "Off"
-        except Exception:
-            self.status = "Error"
-
-    def get_icon(self):
-        WIFI_RUNNING = "直"
-        WIFI_INIIT = "睊"
-        WIFI_NG = "Error"
-
-        if self.status == "init":
-            return WIFI_INIIT
-        elif self.status == "Off":
-            return WIFI_INIIT
-        elif self.status == "running":
-            return WIFI_RUNNING
-        else:
-            return WIFI_NG
-
-
-if __name__ == "__main__":
-    ssid_now = ssid()
-    print(
-        "#[fg={}]".format("colour045" if ssid_now.status == "running" else "colour1")
-        + ssid_now.get_icon()  # noqa: W503
-        + "#[default] "  # noqa: W503
-    )
