--- tmux/window/git-window
@@ -1,1 +1,0 @@
-new-window -n git lazygit
