--- tmux/window/nvim-window
@@ -1,1 +1,0 @@
-new-window -n nvim nvim
