--- tools/alacritty/setup.sh
@@ -1,2 +1,0 @@
-#/bin/bash
-curl -LO --output-dir ~/.config/alacritty https://github.com/catppuccin/alacritty/raw/main/catppuccin-mocha.toml
