--- tools/asdf/.tool-versions
@@ -1,2 +1,0 @@
-nodejs 22.8.0
-ruby 3.3.4
