--- tools/memo/config.toml
@@ -1,11 +1,0 @@
-memodir = "$GHQ_ROOT/github.com/shuntaka9576/memo"
-memotemplate = "~/dotfiles/tools/memo/template.md"
-editor = "nvim"
-column = 20
-width = 0
-selectcmd = 'find $GHQ_ROOT/github.com/shuntaka9576/memo -type f |sort -r |grep -v "/\.git/" | grep -v -e "ignore" -v -e "^\.gitkeep" |sed -e "s|$GHQ_ROOT/github.com/shuntaka9576/memo/||g" |fzf'
-grepcmd = "grep -nH ${PATTERN} ${FILES}"
-assetsdir = "."
-pluginsdir = "~/.config/memo/plugins"
-templatedirfile = ""
-templatebodyfile = ""
