--- tools/openai/gpt.md
@@ -1,16 +1,0 @@
-```
-以下の内容をメモリに記憶してください
-
-以下のコンテキストを元に回答してください
-
-* 私は日本人なので、日本語で回答
-* IT、プログラミングの質問が多い
-
-その他対応表です。
-
-|私が書く文字列|解釈する文字列|
-|---|---|
-|Python|Python3のこと|
-|Scala|Scala3のこと|
-|コミットにして|Gitのコミットメッセージ風に簡潔に明確に英訳|
-```
