--- tools/pet/config.toml
@@ -1,22 +1,0 @@
-[General]
-  snippetfile = "~/.config/pet/snippet.toml"
-  editor = "sensible-editor"
-  column = 40
-  selectcmd = "fzf-tmux --reverse --prompt=\"Snippet > \""
-  backend = "gist"
-  sortby = ""
-
-[Gist]
-  file_name = "pet-snippet.toml"
-  access_token = ""
-  gist_id = ""
-  public = false
-  auto_sync = false
-
-[GitLab]
-  file_name = "pet-snippet.toml"
-  access_token = ""
-  url = ""
-  id = ""
-  visibility = "private"
-  auto_sync = false
