--- tools/testdata/snippet_1.json
@@ -1,22 +1,0 @@
-{
-  "data": {
-    "history": {
-      "list": [
-        {
-          "count": 1,
-          "user": {
-            "id": "A",
-            "displayName": "foo"
-          }
-        },
-        {
-          "count": 2,
-          "user": {
-            "id": "B",
-            "displayName": "foo2"
-          }
-        }
-      ]
-    }
-  }
-}
