--- zsh/.zlogin
@@ -1,3 +1,0 @@
-#!/usr/bin/zsh
-
-source $ZDOTDIR/.zplug/repos/shuntaka9576/prezto/runcoms/zlogin
